requirements.txt
# Dependencias principales del portafolio
Flask==2.3.3
Flask-CORS==4.0.0
PyJWT==2.8.0

# Base de datos
sqlite3

# Utilidades para desarrollo web
Werkzeug==2.3.7
Jinja2==3.1.2

# Para manejo de fechas y tiempo
python-dateutil==2.8.2

# Para validación de datos
email-validator==2.0.0

# Para encriptación de contraseñas
bcrypt==4.0.1

# Para manejo de archivos CSV/Excel (cuando se necesite)
# pandas==2.1.1  # Comentado por problemas de compilación en Windows
openpyxl==3.1.2

# Para desarrollo y testing
pytest==7.4.2
pytest-flask==1.2.0

# Para documentación de API
flask-restx==1.1.0

# Para variables de entorno
python-dotenv==1.0.0